+++ Changelog.md
@@ -3,6 +3,7 @@
 ## 1.3.0
 - Fixed FormatException in translation
 - Fixed output bug by locking the console output
+- Added script line numbers to scriban runtime exceptions
 
 ## 1.2.5
 - Compiler debug information fix